Allow configs to interchange meta or qty
A problem was found where even in the default config, minecraft:gold_nugget*15 was included but did not work.

This is because ConfigParser considers the split between non-alpha characters  to be of form modid:meta if the size == 3.

This proposed change allows the user to use either -- if there's an '*' in the line, then it's assumed to be qty, otherwise it's assumed to be metadata. Therefore modid:item:meta and modid:item*qty both work independently.

This  closes #51